Centralize requirements and use whole plugins and DAG folder to run cloud functions
