+++ dags/test_dag.py
@@ -1,8 +1,6 @@
 from limber.models.dag import DAG
 from limber.operators.python_operator import PythonOperator
-from datetime import timedelta
-import pandas
-
+from plugins.test_utils import test, test_multiple_outputs, test_with_context
 
 default_args = {
 
@@ -15,15 +13,6 @@ test_dag = DAG(
     schedule_interval="0 * * * *"
 )
 
-
-def test(arg):
-    import pandas as pd
-    from datetime import datetime
-
-    df = pd.DataFrame([[datetime.now(), arg]], columns=["current_timestamp", "message"])
-    df.to_gbq("test_dataset.test_table", if_exists="append")
-
-
 test_task = PythonOperator(
     dag=test_dag,
     task_id="test_task",
@@ -33,12 +22,6 @@ test_task = PythonOperator(
     requirements=["pandas==1.0.0", "pandas_gbq==0.14.0"]
 )
 
-
-# Use an array as output to call task 3 two times
-def test_multiple_outputs(arg):
-    return ["test1", "test2"]
-
-
 test_task2 = PythonOperator(
     dag=test_dag,
     task_id="test_task2",
@@ -47,15 +30,6 @@ test_task2 = PythonOperator(
     op_kwargs={'arg': 'great test2'},
     requirements=["pandas==1.0.0", "pandas_gbq==0.14.0"]
 )
-
-def test_with_context(arg, context):
-    import pandas as pd
-    from datetime import datetime
-
-    print(context)
-
-    df = pd.DataFrame([[datetime.now(), arg]], columns=["current_timestamp", "message"])
-    df.to_gbq("test_dataset.test_table", if_exists="append")
 
 test_task3 = PythonOperator(
     dag=test_dag,
