+++ requirements.txt
@@ -1,0 +1,2 @@
+pandas==1.0.0
+pandas_gbq==0.14.0
